+++ features/transaction/checkout/src/main/java/com/tokopedia/checkout/data/model/request/checkout/ShopProductCheckoutRequest.kt
@@ -10,7 +10,7 @@ import java.util.*
 data class ShopProductCheckoutRequest(
         @SuppressLint("Invalid Data Type")
         @SerializedName("shop_id")
-        var shopId: Int = 0,
+        var shopId: Long = 0,
         @SerializedName("is_preorder")
         var isPreorder: Int = 0,
         @SerializedName("finsurance")
